账号表按邮箱哈希切成 64 个分区，计数全面封顶
十亿账号塞进一张表，问题不是查询慢，是没法维护：VACUUM 要跑几十小时，
建一次索引锁上半天，删一批账号产生的死元组够 autovacuum 追一整天，
备份与恢复没有任何粒度。

分片键选邮箱而不是 id 有个硬理由。PostgreSQL 的分区表只能建包含分区键的
唯一约束，按 id 分片就保不住 email 全局唯一，而查重是导入的地基；按邮箱
哈希分片时 UNIQUE (shard, email) 恰好等价于 UNIQUE (email)——同一个邮箱
必然算出同一个分片，不可能出现在两个分区里。代价是按 id 查要在 64 个分区
上各探一次索引（不到 1 毫秒），换来的是按邮箱查——取件 API 最热的路径——
直接裁剪到六十四分之一。

切分区要整表重写并全程持锁，因此超过一百万行就只记日志不动手，下次启动
再判断。那种代价必须由人挑时间承担，而不是某次例行重启时冷不丁发生。
SQLite 不分区也不需要，shard 与 domain 两列照样写，两边共用同一套 SQL。

计数是另一处不设防的地方。总览页与调度器原来有五条 COUNT(*)，十万行
146 毫秒，线性外推到十亿行是二十多分钟——页面直接超时，而调度器每轮都要
读它。改成扫够就停的封顶计数，配上字面匹配的部分索引，代价与结果规模成
正比而不是与表规模成正比。封顶时如实标出来，界面显示"10 万+"，不把一个
下限当精确值给人看。总数改用 PostgreSQL 自己维护的行数估计，十万以内仍
数准——导入一百个账号却看到"约 98 个"像是系统丢了东西。

取任务也从"一条查询算优先级再全局排序"拆成四档分别取。十亿账号每天到期
一千六百万行，原来的写法等于每分钟为了拿 30 条任务排序一千六百万行。

顺带补齐三处一直缺索引的路径：领取空闲账号（原来是全表扫描加临时排序）、
按域名筛选（原来是 LIKE '%@域名'，前缀通配用不上任何索引）、按 id 查账号。
域名单独存了一列，域名聚合也不必再在 SQL 里现切邮箱后缀。

分区是 PostgreSQL 独有的，本地 SQLite 测不出来，因此补回 CI 测试工作流，
用真实 PostgreSQL 跑一遍全部用例，包括从有数据的普通表升级成分区表这条
路径——数据搬丢了不会有人当场发现，而序列没续上会让新账号拿到用过的 id。

+++ backend/internal/scheduler/autorate_test.go
@@ -20,7 +20,7 @@ func TestDeriveRatesScalesDown(t *testing.T) {
 
 // 速率不能为 0：那意味着调度器完全不工作，账号会一直等到过期。
 func TestDeriveRatesNeverZero(t *testing.T) {
-	for _, n := range []int{0, 1, 10} {
+	for _, n := range []int64{0, 1, 10} {
 		d := DeriveRates(n, 60, 1, 1)
 		if d.PerIPPerMin < 1 || d.PerClientPerMin < 1 {
 			t.Fatalf("账号数 %d 推出速率 %d/%d，不能低于 1", n, d.PerIPPerMin, d.PerClientPerMin)
